refactor(budget): dedupe sanitize_* + FakeProviderQuotaSnapshots
Closes #854
