+++ directory-tree.md
@@ -187,7 +187,9 @@ maestro/
 │   ├── continuous.rs                      # ContinuousModeState and ContinuousFailure structs; state machine for --continuous / -C flag: auto-advances to next ready issue, pauses loop on failure waiting for user decision (skip / retry / quit)  [Issue #85]
 │   ├── budget.rs                          # BudgetEnforcer: per-session and global budget checks; `mod projector`, `mod quota_snapshot` declared; `check_pre_spawn()` + `PreSpawnDecision` added  [Phase 3, Issue #848]
 │   │   ├── projector.rs                   # Budget pre-spawn projector — pure fns; projects remaining headroom before a session starts  [Issue #848]
-│   │   └── quota_snapshot.rs              # `QuotaSnapshot` trait — provider-agnostic read interface for quota state; implemented by MiniMax quota  [Issue #848]
+│   │   ├── quota_snapshot.rs              # `QuotaSnapshot` trait — provider-agnostic read interface for quota state; implemented by MiniMax quota  [Issue #848]
+│   │   ├── sanitize.rs                    # Value-clamp primitives shared between budget projector and TUI rollup; deduped from inline clamping logic  [Issue #854]
+│   │   └── test_support.rs               # `#[cfg(test)]` shared `FakeProviderQuotaSnapshots` fake — eliminates per-module duplicate fakes  [Issue #854]
 │   ├── tui/budget_prespawn.rs             # Pre-spawn budget gate modal widget: renders [y]es/[n]o/[s]kip confirmation overlay before session spawn when projected spend exceeds threshold  [Issue #850]
 │   ├── tui/budget_banner.rs               # 80% global budget banner widget: top-chrome warning strip rendered when aggregate spend crosses 80% of the configured total budget  [Issue #850]
 │   ├── doctor.rs                          # Preflight checks: CheckSeverity, CheckResult, DoctorReport, run_all_checks(), print_report(); validate_preflight() (public, fails fast on required check failures); build_claude_cli_result() (pub(crate), pure/testable); check_claude_cli() elevated to Required severity; build_gh_auth_result() (pure, testable); check_az_identity(); 10 check functions  [Issue #49, #34, #52]
